requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0

# Environment & Configuration
python-dotenv==1.0.0

# OpenAI Integration
openai==1.6.1

# Vector Database
chromadb==0.4.22

# Embeddings & NLP
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Evaluation & Metrics
scikit-learn==1.3.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
